Add paragraph about RtpContributingSources being updated simultaneously.
Fixes #1124.

The intent of this PR is to ensure that if an application reads the
attributes of multiple `RTCRtpContributingSource`s, the values are all
taken from the same point in time, so they're comparable.

For example, an application may compare timestamps to see which
contributing source(s) are actively contributing (have the most recent
timestamp), or see which one is loudest. Without the guarantee this PR
attempts to add, a browser might return values from different points in
time for different `RTCRtpContributingSource`s, making such comparisons
unreliable.